+++ enigma/utils/database.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 from datetime import date, datetime as d
-from typing import List, Optional
+from typing import List, Optional, Union, Tuple
 
-from sqlalchemy import create_engine, Column, Integer, Date, BigInteger, Text
+from sqlalchemy import create_engine, Column, Integer, Date, BigInteger, Text, text
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
@@ -14,6 +14,14 @@ _Session = sessionmaker()
 _Session.configure(bind=_engine)
 
 Base = declarative_base()
+
+
+def check_connection() -> bool:
+    session = _Session()
+    try:
+        return session.is_active
+    finally:
+        session.close()
 
 
 class User(Base):